♻️ adjustment in delete teacher

+++ app/modules/Teacher/repositories/DeleteTeacherRepository.ts
@@ -4,16 +4,18 @@ import { TDeleteTeacher } from '../type'
 import { Exception } from '@adonisjs/core/build/standalone'
 
 export class DeleteTeacherRepository {
-  public async handle({ ctx, secureId }: TDeleteTeacher) {
+  public async handle({ secureId }: TDeleteTeacher) {
     try {
-      const teacher = await Teacher.query().where('secure_id', secureId).first()
-
+      const teacher = await Teacher.query().where('secure_id', secureId).preload('course').first()
       if (!teacher) {
         throw new NotFoundException('There is no teacher for this secure id', 404, 'E_NOT_FOUND')
       }
+      console.log(teacher?.toJSON())
 
-      await teacher.delete()
-      return ctx.response.ok({ message: 'Teacher deleted successfully' })
+      if (teacher?.course.length > 0) {
+        throw new NotFoundException('this teacher is linked to a course', 422, 'E_USER_LINKED')
+      }
+      return teacher.delete()
     } catch (error) {
       throw new Exception(
         error.message || 'Internal Server Error',
